fixing endpoint

+++ backend/app/api/esami.py
@@ -4,9 +4,11 @@ from flask import jsonify, request
 from flask_jwt_extended import *
 from sqlalchemy import insert
 
+from app.utils import genera_date_appello
 from app.models.Esame import Esame
 from app.models.Prova import Prova
 from app.models.Realizza import Realizza
+from app.models.Appello import Appello
 
 @bp.route('/esami/crea', methods=['POST'])
 @jwt_required()
@@ -22,6 +24,9 @@ def inserisci_in_libretto():
         anno = request.json["anno"]
         prove = request.json["prove"]
         collaboratori = request.json["collaboratori"]
+ 
+        if not idesame or not nome or not crediti or not anno or not prove or not collaboratori:
+            return jsonify({"Status": 401, "Reason":"Missing parameters!"})
 
         query = insert(Esame).values(
             idesame = idesame,
@@ -41,14 +46,24 @@ def inserisci_in_libretto():
 
         #TO-DO: opzionale deve avere valori uguali a true o false
         for index, prova in enumerate(prove, start=1):
+            idprova = f"{idesame}-{index}",
             query = insert(Prova).values(
-                idprova = f"{idesame}-{index}",
+                idprova = idprova,
                 tipologia = prova.get("tipologia"),
                 opzionale = prova.get("opzionale"),
                 datascadenza = prova.get("datascadenza"),
                 dipendeda = prova.get("dipendeda") if prova.get("dipendeda") != "" else None,
                 idesame = idesame,
                 responsabile = current_user['email']
+            )
+            session.execute(query)
+            
+        date_appelli = genera_date_appello(4)
+
+        for data in date_appelli:
+            query = insert(Appello).values(
+                idprova = idprova,
+                data = data
             )
             session.execute(query)
 
@@ -78,9 +93,13 @@ def elimina_esame():
 
         idesame = request.json["idesame"]
 
+        if not idesame:
+            return jsonify({"Status": 401, "Reason":"Missing parameters!"})
+
         session.query(Esame).filter(Esame.idesame == idesame).delete()
         session.commit()
         return jsonify({"Status": "Success"}),200
     except:
         session.rollback()
         return jsonify({"Status": "Failure"}), 500
+
